rename: 폴더명 profile -> signup 으로 변경 #153
